I went through and added a rudimentary GUI for the Server and for the Client. The Server can now choose its own port and start or stop the server at will. This seems to lag 1 cycle though, which you can figure out if you test it with clients.
The Client allows you to specify the portnumber and the ip address of the server.
The server requires 2 players, so it will wait until it receives another client.
So you should spawn two Client instances to test it out.

All the server does right now is send a customized message to each player to show that it is working.
I intend on building a framework/some infrastructure for displaying and moving sprites.
From there Sean and I can work on creating menus, dialogues, a title screen, and other such things.

I am thinking we can have multiple scenes, and we can change the scene once the settings are set.
So, once you connect to the server the app will automatically switch to the game screen/scene, etc.

~Logan