I was looking into how to make the GridPane work with moving ships around, and displaying a background and a ship over it, but I couldn't find a way to overwrite the rendering to make it custom. Frustrated, I coded a demo using Canvas and a GridPane, and then I wasn't satisfied stopping there so I spent many many hours getting Client Server communication working.
Right now I have a simple game demo working. Click on your piece, then click 3 squares away to move it. If an enemy is in-range, clicking them will kill them and end your turn. The tiles turn gray when your turn is over. I use System.out.println() to send messages to the console, so if you click a ship and try and move it and it's not working, then check the console to see what is the issue.

I want to add buttons and textareas to the gridpane next, but that requires more math and I haven't drawn it out yet. Let me know what you think, I can demo it in person if you guys are free to meet sometime.

This is just a demo, we don't have to use a Canvas element. I just wanted to demo what a solution using Canvas might look like, and also wanted something that works as a foundation we could build upon.

I intend on replacing the ships with images, and the tiles with randomized space-like images, such as asteroids or something.

Anyway, lemme know what you think. It's a bit bland, but it's cool I think.